vision/stages/icon-classification/pyproject.toml
[project]
name = "icon-classification"
version = "0.1.0"
description = "Per-button icon classification stage (shape heuristic + optional CoreML/ONNX model)"
requires-python = ">=3.11"
dependencies = [
    "testanyware-common",
    "numpy>=1.26",
    "pillow>=10.0",
]

[project.optional-dependencies]
onnx = ["onnxruntime>=1.18"]
# CoreML requires macOS-only coremltools; leave as an opt-in.
coreml = ["coremltools>=7.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/icon_classification"]
